actions: Add buffering to the channels
This is because if the Run-function e.g. for ShAction is handling the
startchan operation, then the sending a message to the termchan will hang in
the Receive-function.

Adding a buffer puts the terminate-message to the buffer and it is handled
just after the Run is not busy any longer.